Forgoten change when switching to Integer from long in some cases.

+++ src/ghzclicker/Controller.java
@@ -139,7 +139,7 @@ public class Controller {
 		for (int i = 0; i < gui.getBtnBuildings().size(); i++) {
 			hertzPerSecond += buildings.get(i).getOwned() * buildings.get(i).getBaseHPS();
 		}
-		gui.updateHertzPerSecond(Long.toString(hertzPerSecond));
+		gui.updateHertzPerSecond(Integer.toString(hertzPerSecond));
 	}
 
 	/**
